Updated Auction detail screen
